Fix bug in config generation algorithm
The algorithm failed to find valid configuration for syntethizable
frequencies, for example 130 MHz, the OMD integer was set to
an incorrect value.

This happened because 6 was not considered a valid value for the OMD
integer (according to AN-619), however Clock Builder uses integer
and it seems its the only way to syntethize frequencies at this range.

k